Add DB for AdNetworks, mark III
Binary files apparently present a significant issue in Rietveld, Git, SVN, and
life on the whole. In order to counteract this, kalman@ and I decided to go with
a hash table of strings in a .cc file (this was after a number of other attempts
which ended badly to due endianness and other mishaps). This is a hair larger,
but is platform independent, and should actually be able to be committed.

Also included: A handy-dandy python script to generate the lists from a given
input file.

We had to abandon the base::hash_set<int64> method, as cool as it was, because of byte ordering issues, which we could not comfortably manage when transitioning between as many platforms as this has to run on.  Instead, we use a c-style array, binary search, and strcmp, which should still result in a very fast lookup time (though not quite as elegant).

BUG=357204

Review URL: https://codereview.chromium.org/277453004