Support input events in Chrome ASH on Windows 7.
All changes are in the metro_driver_win7.cc file. There are two classes MouseEvent and KeyboardEvent
which implement the required set of winrt interfaces needed for this to work. We call the corresponding
handlers in the ChromeAppViewAsh class on the same lines as ASH on Windows 8.

BUG=356475
R=cpu@chromium.org, cpu

Review URL: https://codereview.chromium.org/282953003