Move a good set of gfx unit tests into gfx_unittests target.
This requires to major changes:

1- It requires the introduction of GfxTestSuite to allow us to
initialize the ResourceBundle which is required for some of these tests
to pass.
2- It requires to split gfx_unittests target out of gfx.gyp, into
gfx_tests.gyp. This is necessary because otherwise a circular dependency
between ui_base.gyp and gfx.gyp would arise.

The circular dependency is the form of:
gfx.gyp -> ui_base.gyp -> gfx.gyp

And is created because gfx_unittests now requires ResourceBundle to link
which is in ui_base target.

BUG=331829
TEST=ui_unittests,gfx_unittests
R=danakj@chromium.org
TBR=ben@chromium.org

Review URL: https://codereview.chromium.org/280973003