Refactor ViewsTestBase to isolate aura dependency.
We want to run views_unittests without aura for the mac port of views. 
This is accomplished by introducing a ViewsTestHelper interface to abstract 
the platform-dependent parts of test setup and teardown. It is responsible for 
creating and cleaning up the platform specific objects that are needed for
running tests.

BUG=366007

Review URL: https://codereview.chromium.org/287683002