Files.app: Add an API to resolve isolated file system to a file entry in the local file system.
Entries passed via FileHandler are contained in the isolated file system, so we
cannot obtain their parent directory or other entries in the same directory.

This CL adds an API to resolve isolated file system to a file entry in the local
file system.

BUG=358698
TEST=manually

Review URL: https://codereview.chromium.org/279163002