Remove hit-testing-only RSLL entries
As of https://codereview.chromium.org/266913021/ we no longer hit test
via the RSLL; we walk the tree instead. This means we can remove the
machinery added to allow hidden layers to appear in the RSLL for hit
testing purposes.

This is essentially a revert of https://codereview.chromium.org/25712004

R=danakj@chromium.org
BUG=None

Review URL: https://codereview.chromium.org/289973003