Add a failure expectation to win memory.css3d test.
In tile manager we seem to reach the memory limit early (with the
pending tree). However, when we activate our memory gets released
and we start filling it up again with the now active tree tiles.

The windows bot seems to catch the system at the moment when we're not
using a lot of memory, thus failing the test.

BUG=373098
R=kbr@chromium.org

Review URL: https://codereview.chromium.org/289003004