Domain Reliability: Don't use error code unless request failed
I was copying the error code out of the URLRequestStatus of a monitored
request whether or not the status was FAILED; in cases where it wasn't,
the error would be something random (often QUIC_PROTOCOL_ERROR).

This change makes sure that we only copy the error code when we should,
and adds a unit test to make sure that beacons for successful requests
contain net::OK even if the error field contains something else.

BUG=356791

Review URL: https://codereview.chromium.org/284103003