ash: Do a null-check for the screen-position client.
In a multi-monitor setup, when switching from extended mode to mirror mode,
it is possible for the external monitor to receive events between when its
RootWindowController is Shutdown()ed (which destroyes the screen-position
client) and the root-window is destroyed. So make sure the client is there
before trying to use it to convert the event's location to screen space.

BUG=358266
R=oshima@chromium.org

Review URL: https://codereview.chromium.org/286913009