ozone: Add ozone_auto_platforms setting
Each time we add a new platform in the default build, products that want
only a single platform have to add an entry to GYP_DEFINES to turn it back
off.

This introduces a new setting, ozone_auto_platforms. If this is turned off
in GYP_DEFINES, we'll only build platforms that are explicitly enabled.

TEST=runhooks
BUG=none

Review URL: https://codereview.chromium.org/288273002