Use GUID instead of ServicePath in networkingPrivate API
Currently we pass the service path of networks as the "GUID" in the
networkingPrivate implementation on Chrome OS. We should be using the
actual GUID (if one exists) or create a GUID if one does not.

Note: This change should not break any existing usage since the GUID
value should be transparant to the implementation. (This will have the
positive side effect of making the GUID consistent across sessions).

BUG=284827
For fake_wifi_service.cc:

R=asargent@chromium.org, pneubeck@chromium.org
TBR=mef@chromium.org

Review URL: https://codereview.chromium.org/275543005