Revert 264676 "Make sure the ClearBrowserDataOverlay is disabled..."
Note that this revert is only performed on release branch 1985 (M36). The reason is that r264676 introduced 3 regressions that require a more complex CL to fix, and we are not comfortable merging that at this stage. On trunk, r264676 will not be reverted, and a follow-up CL will address the regressions.

> Make sure the ClearBrowserDataOverlay is disabled as long as a prior clearing is still running.
>
> See the bug for details.
>
> BUG=354931
>
> Review URL: https://codereview.chromium.org/215193004

BUG=365206
TBR=engedy@chromium.org

Review URL: https://codereview.chromium.org/290573012