[dom_distiller] Fix crash for invalid chrome-distiller:// requests
When the DomDistillerViewerSource got a request for chrome-distiller://foobar/
it would crash if there was an empty path argument. This CL adds a safety check
for the length of the path, to ensure to not take a substring of an empty
string. 

In addition, this adds regression tests for the crash it fixes. It also fixes
the rest of the disabled browser tests for the DomDistillerViewerSource. 

BUG=356866

Review URL: https://codereview.chromium.org/288353002