Make sure ReadyStateEnded is fired when video stream stops
A regression was introduced in the refactoring change in r269020.
VideoCaptureDelegate::started_callback_ is used to notify that the
stream has started and ran into error. However in change r269020 the
callback was reset after sending the started signal. This prevented
error signal from firing.

In this change |started_callback_| is renamed to |running_callback| to
avoid confusion. Also added a test to verify that ReadyStateEnded is
reached when a camera stops.

BUG=374495

Review URL: https://codereview.chromium.org/295483002