Add test data for dynamic CLD.
This is necessary in order to allow browser tests to run correctly when using
CLD in dynamic mode (either from an external file or via the component
updater). This change is being split apart from the actual tests that use it
because the CLD data file is too large to commit via the CQ and, as a result,
must be dcommit'ed.

BUG=367239
R=rmcilroy@chromium.org

Review URL: https://codereview.chromium.org/289313004