Fix check preventing gsutil from being included in standalone telemetry.
We were checking if 'prodaccess' is in the path when trying to decide
if 'gsutil' supported prodaccess; when we called this as part of
the find_dependencies script, this meant that we wouldn't include gsutil
because prodaccess wasn't on the bot.

We really only want the path check to happen when we're handling an
authorization failure, in order to determine what message to show the
user.

This CL fixes this issue.

R=nduca@chromium.org, dtu@chromium.org
BUG=369350

Review URL: https://codereview.chromium.org/291293002