Save running SW instance info, including its SiteInstance, into the ProcessManager.
We can use this map to drop all the process references on shutdown, in order to
fix http://crbug.com/368570.

This starts to fix an assumption in the content layer that SiteInstances outlive
their RPHs so that process permissions can be controlled through the
SiteInstance instead of by process id. However, there's still work to do before
that's completely fixed.

BUG=368570

Review URL: https://codereview.chromium.org/292903002