Do not call SendMessageToNativeLog on Android in AudioInputSyncWriter
MediaStreamManager::SendMessageToNativeLog posts a task to the UI thread, which will attach the audio thread to the Android java VM. Unlike chrome created threads, the audio thread is owned by the OS and does not detach itself from the VM on exit, causing a crash.

The fix is to not call SendMessageToNativeLog on Android because webrtc logging is never enabled on Android due to the lack of extension.

BUG=365915

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=271764

Review URL: https://codereview.chromium.org/294803003