Always write .landmines
Otherwise, if .landmines exists, and isn't clobbered by the clobber in
compile.py, every build will cause a clobber until .landmines is removed.

Trying to fix http://crbug.com/375459 brought this up. In that case
compile.py wasn't removing .landmines, which is necessary in the case
of a landmines change on two subsequent builds. If .landmines isn't
maintained, then there's nothing to diff against on the second build
and so a clobber won't be correctly triggered.

R=iannucci@chromium.org,kbr@chromium.org
BUG=376359

Review URL: https://codereview.chromium.org/293983017