Synthetic pinch gesture take scale factor as parameter.
Instead of a zoom direction and the number of pixels to be covered by the pinch
gesture, the gesture now takes a scale factor as parameter. From that, it
computes how to perform the gesture: the ratio of the final span and the initial
span equals the scale factor (modulo touch slop). The spans must not fall below
a certain threshold, provided by the gesture target (GetMinScalingSpanInDips).

Update the JS interface (chrome.gpuBenchmarking.pinchBy) to reflect the changes.
For (temporary) backwards compatibility, add another method to the interface to
tell Telemetry whether or not the browser supports the new interface. This is to
avoid crashes when Telemetry is run with the current stable version. Will be
removed once the stable version supports the new interface.

Update tough_pinch_zoom_cases page set to make sure all pinch gestures start
within the web content (tested on N4 and N10).

Update synthetic gesture controller unit tests to use the new parameter.

BUG=309484

Review URL: https://codereview.chromium.org/293683002