There can be potentially unexpected message names in the bit-flip test.
This removes some NOTREACHED()s in the base class which may be
reached when bad data comes along.  Merely tally such so that it
can be checked in the cases where we know it's "impossible".

We also need to be more lenient about the results of reading
from pipes.

BUG=376173

Review URL: https://codereview.chromium.org/297833007