Early return in UpdateTouchTransformer for software mirroring mode
Software mirroring mode is special that display_state is marked as
DUAL_EXTENDED but it is doing mirroring. There is only one root
window as supposed 2 root windows in the extended mode. Early return
in UpdateTouchTransformer() for the software mirroring mode for now
so as not to crash chrome.

BUG=376459
TEST=arm devices don't crash while switching from extended to mirror
mode.

Review URL: https://codereview.chromium.org/287193014