Clear EmbeddedWorkerContextClient's thread-local ptr in willDestroyWorkerContext()
Back then EmbeddedWorkerContextClient used to be destroyed on the worker
thread when corresponding WorkerGlobalScope's destructed, but after
http://crrev.com/262813004 we started to destruct it on main thread
(after the worker side termination), so resetting thread-local ptr in
the dtor doesn't work (as it runs on main thread).

This change clears the thread-local ptr in willDestroyWorkerContext(),
where we're sure we're still on the worker thread.

BUG=n/a

Review URL: https://codereview.chromium.org/287193011