Fetch a new GAIA picture every time chrome://settings/manageProfile comes up.
(currently we do it once, when a profile is signed in, and henceforth every 24h)

This assumes the --google-profile-info flag is in use (which comes for free
as part of --new-profile-management), and is similar to CrOS, which also forces
an update every time chrome://settings/changePicture is loaded.

I've also deleted the code that forced the "is using the gaia photo" flag to true
on every completed download, as it meant that we'd overwrite any custom avatars
the user may have chosen.

BUG=311013

Review URL: https://codereview.chromium.org/267253004