Make the App Info Dialog appear modal
Gave the App Info Dialog a 10px 'gap' around the edges (so
the app list is visible through it), and a larger gap in
the wide app list so it appears as a vertical column down
the middle.

Also made a white background overlay that appears/disappears
with the dialog.

BUG=369888,374080,374076

Review URL: https://codereview.chromium.org/276833002