Delete EmbeddedWorkerDispatcher in RenderThreadImpl::Shutdown().
In current code, EmbeddedWorkerDispatcher is deleted in the destructor of RenderThreadImpl.
If EmbeddedWorkerDispatcher contains any WorkerWrapper in workers_, WorkerWrapper and EmbeddedWorkerDevToolsAgent will be deleted at that time.
And the destructor of EmbeddedWorkerDevToolsAgent will call RemoveEmbeddedWorkerRoute of RenderThreadImpl which is being deleted.

To avoid this we should delete EmbeddedWorkerDispatcher in RenderThreadImpl::Shutdown().

BUG=377258,358657

Review URL: https://codereview.chromium.org/297343002