Experimental app list slide animation is now horizontal, not vertical.
This makes the lack of proper clipping of WebContents more noticeable,
but it's behind a flag so fixing it isn't urgent.

Note that the animation always slides pages to/from the left side
(previously the top). In the future, the pages will be positioned in the
correct order.

BUG=376068

Review URL: https://codereview.chromium.org/305433002