Show logs in english in TracingController for adb_chrome_profiler
if the target is not in english, adb_profiler_chrome can 
not know when the profiler is started and finished.
because adb_profiler_chrome uses only english regex for
parsing the logs. So this is a patch for showing the logs 
in english, however it still shows toasts by following 
the target's locale.

BUG=373505

Review URL: https://codereview.chromium.org/282813002