Fix a bug in gpu_unittests mock expectations for VAO.
I notcied this bug while trying to fix the WebGL VAO related regressions
on Linux AMD.

Before the generated ID is not set, so by default it's 0, which
is wrong, because 0 is reserved to unbind instead of bind.

Also, turn on VAO support for majority of tests as the emulated code path
is more isolated and its tests are also more isolated.  I feel it's better
to test with VAO on by default.

BUG=373974
TEST=gpu_unittests
R=vmiura@chromium.org,kbr@chromium.org,piman@chromium.org,jbauman@chromium.org

Review URL: https://codereview.chromium.org/303543005