Battery Status API: browser-side implementation.
This patch implements the browser side bits needed for the
Battery Status API. It includes a message filter, a singleton
service which owns and controls the platform-specific battery
status manager. Updates are implemented using callbacks and
async IPCs between the browser and the renderer.

A unit test for the service using a fake battery status manager
and callbacks is also provided.

Corresponding renderer-side patch:
https://codereview.chromium.org/252113006/

BUG=122593, 360068

Review URL: https://codereview.chromium.org/278043002