Add a blank interstitial page for HTTP Auth prompt.
This patch displays a blank interstitial when HTTP basic
auth login prompt is triggered. The interstitial effectively 
clears the page and forces the omnibox to show the correct
URL for the auth request. It's only shown for cross origin
requests in the main frame.

BUG=149871
TEST=LoginPromptBrowserTest.ShowCorrectUrlForCrossOriginMainFrameRequests

Review URL: https://codereview.chromium.org/276123002