ui: Fix ui_unittests assertion on iOS.
In order to iOS to find the test data files, we need to copy them into
the device. To do this we need the help of
build/copy_test_data_ios.gypi.

This patch includes it in the ui_unittests target when the platform is
iOS and fix the TODO in data_pack_unittests.

By doing that we were able to ASSERT_TRUE the PathService::Get() call.

BUG=None
TEST=ui_unittests on iOS bots.
R=tony@chromium.org

Review URL: https://codereview.chromium.org/297183005