Cast: Synthetic benchmark tool.
This benchmark tests how much latency, packet loss and bandwidth loss we can tolerate.
It runs hundreds of times faster than real-time and results are independent of CPU speed.
So far, it does a series of binary searches to find a spanning tree of the space in which
our algorithms can operate comfortably.

Review URL: https://codereview.chromium.org/293173002