Show window even if the parent window already contains the window
The window may have changed visibility in the meantime. The web contents
controller may have navigated back, thus the window becomes visible. In this
case the window needs to be shown even if the parent contains it.

This aligns the shell_aura.cc implementation with shell_views.cc since in
the Views implementation the View::AddChildView() call would perform this
check and the widget would be shown by the next call to resize the widget
which happen regardless if View::AddChildView() was a no-op.

BUG=none

Review URL: https://codereview.chromium.org/308463002