Remove webrtc_internals log spam
I see "Could not get the download directory" in a lot of test output.
This is not relevant to any of these tests, so this log should just
become a vlog that can be turned on via command line by developers who
care about it.

R=vrk@chromium.org
BUG=none

Review URL: https://codereview.chromium.org/308673003