Add human readable programmatic enum name/values to chrome.automation.
This adds three new bindings to the automation API:
chrome.automation.Event
chrome.automation.Role
chrome.automation.State

Each is an object mapping an enum name to its value. Currently, this is an identity mapping. However, with the way the cl constructs the bindings, we could in the future substitute with integer values (and eliminate the stringification on either end of our extension/browser IPC).

BUG=308003003

Review URL: https://codereview.chromium.org/304293002