Use a data url in DidStopLoadingDetailsWithPending test.
This is to make the test continue passing after http://crrev.com/181493007
lands. The test was relying on the renderer stopping all loaders when a
new navigation is issued (which would result in a DidStopLoading callback).
After the mentioned CL lands the renderer will not stop+restart loading when
asked to navigate during a load breaking the test.

The fix is to use a data URL which will result in a DidStopLoading immediately
after the didCommitProvisionalLoad callback. Since the test is only checking
whether the UI-side state is correct it doesn't matter that the DidStopLoading
IPC is pending when didCommitProvisionalLoad is being received.

BUG=325351
TEST=content_browsertests

Review URL: https://codereview.chromium.org/306893002