Eliminate dependence of GoogleURLTracker et al. on InfoBarService
This CL changes the dependence of GoogleURLTracker(InfoBarDelegate, MapEntry)
on InfoBarService to instead be a dependence on infobars::InfoBarManager. The
only reason that InfoBarManager wasn't sufficient was that
GoogleURLTrackerInfoBarDelegate obtained the WebContents from the
InfoBarService. This flow is replaced by introducing an OpenURL() API on
GoogleURLTrackerNavigationHelper (GoogleURLTrackerNavigationHelperImpl
implements this API by calling through to its WebContents instance, which is
the same instance associated with the InfoBarService).

BUG=373231,373233

Review URL: https://codereview.chromium.org/293503003