Make DefaultDeleter for Video{De|En}codeAccelerator
This makes a specialization of base::DefaultDeleter so that scoped_ptr will
use Destroy() instead of the destructor. Without this, it seems wrong to
hold VideoDecodeAccelerator in a scoped_ptr, given that letting scoped_ptr
delete it via the destructor would be an error.

Also moves the destructors to protected, to prevent mistakes.

BUG=

Review URL: https://codereview.chromium.org/292183011