Pepper: Narrow locking at llc subprocess start.
Currently, subprocess_mu_ is held the entire time the llc subprocess is
starting. This blocks a larger refactor that I'd like to make as part of nexe
loading.

I'd like to change ServiceRuntime::LoadModule() to be asynchronous to simplify
threading behavior on the plugin side when loading nexe modules. However, this
requires that many methods are made asynchronous, including
Plugin::LoadHelperNaClModule().

PnaclTranslateThread will need to use a pattern similar to WaitForSelLdrStart()
to resume execution on a background thread. However, I don't like the idea of
introducing another mutex and condvar to deal with while holding
subprocess_mu_.

This change narrows the time that subprocess_mu_ is held to make that refactor
possible. I'll make a similar change for the ld subprocess, but I wanted to get
this reviewed before proceeding further.

BUG=333950
R=bbudge@chromium.org

Review URL: https://codereview.chromium.org/308013002