content: Always allocate GpuMemoryBuffers on UI thread.
Prior to this change, the browser compositor allocated
buffers on the UI thread while the buffer allocation for
renderers was done on the IO thread.

This moves all allocation to the UI thread, which makes life
easier for GpuMemoryBuffer implementations and avoids
potentially expensive allocation costs on the IO thread.

BUG=

Review URL: https://codereview.chromium.org/302443008