Move TranslateInfoBarDelegate and OptionsMenuModel to the Translate component.
To solve the dependency on chrome_command_ids, this CL removes the
IDC_OPTIONS_TRANSLATE_* constants and instead defines the command IDs
locally in the file, which is an approach already used at several other places in
the code.

To solve the dependency on themes, the icon ID for the translate infobar is no
longer hardcoded, but fetched from the translate client instead.

BUG=371845
TBR=estade, joaodasilva

Review URL: https://codereview.chromium.org/284313008