Pepper: Move "Core" test to content_browsertests.
This test is currently disabled on Linux Aura because it was running too slowly
and caused flakes. I'm hoping that the test will run faster in
content_browsertests, so we won't need to disable the test in that case.

I'm making a commit that only moves this test so it can be easily reverted if
necessary.

BUG=371873
R=dmichael@chromium.org

Review URL: https://codereview.chromium.org/306393002