Add display_unittests target
As a first step to getting a separate target for the display unittests, this CL is adding a simple target with one unittest. All display_unittests currently running under ui_unittests will be migrated under this target when display_unittests is enabled on the main waterfall.

BUG=333413

Review URL: https://codereview.chromium.org/304183010