Created a separate test page for gpu rasterization.
The new test page uses the viewport tag which makes it easy to sample
pixels from the screenshot. Adding the viewport tag to pixel_css3d page
which is also used for pixel tests because it starts to fail due to
unknown reasons on random bots.

BUG=368495

Review URL: https://codereview.chromium.org/302003015