Add ability to constrain dominant color selection to a HSL range.
This CL adds the ability to supply CalculateKMeanColorOfPNG with a lower
and upper HSL bound which will be used to select the dominant color.

This CL also adds CalculateKMeanColorOfPNG with only a PNG param as a
convenience method, using sane defaults and a utility function
IsWithinHSLRange for checking whether a given HSL is within the given
HSL range.

BUG=376613
TBR=benwells@chromium.org, estade@chromium.org

Review URL: https://codereview.chromium.org/289283004