Changing constructor of BufferedDataSource to accept GURL and CORSMode
BufferedDataSource should accept the 'url' and the 'CORS mode' parameters
in the constructor instead via the initialize() method. This patch makes
this change and now only the 'initializeCallBack' is passed to initialize()
method. Also now the WebMediaPlayerImpl directly queries the BufferedDataSource
whether the resource is local or not, instead of finding it out itself.

Review URL: https://codereview.chromium.org/306953005