linux_aura: Theoretical crash for full screen transition.
The only way this crash could occur is if we're trying to layout the
titlebar, the delegate requests showing the icon, but there isn't an
icon view. In such cases, we shouldn't try to set the bounds on a NULL
view.

(This fix is theoretical; I wasn't able to repro the bug.)

BUG=376313

Review URL: https://codereview.chromium.org/310193002