Ack swap buffers after a timeout
In r272469, a timer that would automatically ack swaps after a
quarter-second passed was removed. The purpose of the timer
was to manage the transition between async and sync displays,
but it had an unintentional side-effect of un-blocking the renderer
in the event that the CALayer's display function was never called.

Re-add a timer explicitly for this purpose.

BUG=378445

Review URL: https://codereview.chromium.org/310183002