Enabling main WebRTC browser test in Chromium waterfalls.
Now that we have separated out the tests that require special resources
to the perf browsertest, we can enable the main test case on the
tree-closing waterfall. This will do the same thing as the content
browsertest but run on the full browser rather than content_shell.

BUG=375240

Review URL: https://codereview.chromium.org/313833002