Pepper: Miscellaneous trusted plugin cleanup.
This removes some more unnecessary methods and moves some more UMA reporting
out of the trusted plugin, reducing the total line count.

It's a little awkward that we have to pass the max value of NaClErrorCode to
the new reporting method, but it keeps that enum out of the public NaCl
interface.

BUG=239656
R=dmichael@chromium.org

Review URL: https://codereview.chromium.org/315753003