Support ephemeral ports in DevTools and Telemetry.
In DevTools, if port 0 is requested, query the HttpServer for the port
actually used, and write the port to a well-known location: the file
"DevToolsActivePort" in the user's profile directory. Currently this
handshaking protocol is only implemented for Chromium, not
content_shell, but this is the only scenario in which flakiness has
been observed.

Add a hidden command-line argument to Telemetry's desktop browser
backend which uses this new code path. The bots will specify this
argument. Telemetry must support browsers two releases back, so once
Chrome 37 goes to the Stable channel, the old initialization path will
be removed. Issue 379980 has been filed to remove the old code.

Tested with and without the new --use-devtools-active-port command
line argument with -v. Verified the new code path is taken. Once these
changes are committed, the bots will begin verifying this new code.

BUG=372560
NOTRY=true

Review URL: https://codereview.chromium.org/315503002